Refactor : SocketSender 클래스 도입
SocketSender 클래스를 도입하여 모든 소켓 데이터 전송을 처리하도록 하였습니다.
