+++ src/main/java/nbbang/com/nbbang/domain/chat/controller/MessageController.java
@@ -8,22 +8,19 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import nbbang.com.nbbang.domain.bbangpan.dto.BbangpanResponseDto;
 import nbbang.com.nbbang.domain.chat.domain.Message;
 import nbbang.com.nbbang.domain.chat.dto.ChatMessageImageUploadResponseDto;
 import nbbang.com.nbbang.domain.chat.dto.message.ChatSendRequestDto;
 import nbbang.com.nbbang.domain.chat.dto.message.ChatSendResponseDto;
 import nbbang.com.nbbang.domain.chat.service.MessageService;
-import nbbang.com.nbbang.domain.party.domain.Party;
 import nbbang.com.nbbang.domain.party.service.PartyService;
 import nbbang.com.nbbang.global.error.GlobalErrorResponseMessage;
 import nbbang.com.nbbang.global.error.exception.CustomIllegalArgumentException;
 import nbbang.com.nbbang.global.interceptor.CurrentMember;
 import nbbang.com.nbbang.global.response.DefaultResponse;
+import nbbang.com.nbbang.global.socket.SocketSender;
 import nbbang.com.nbbang.global.response.StatusCode;
-import nbbang.com.nbbang.global.socket.SocketSendDto;
 import org.springframework.http.MediaType;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -42,9 +39,9 @@ import javax.validation.Valid;
 public class MessageController {
 
     private final MessageService messageService;
-    private final SimpMessagingTemplate simpMessagingTemplate;
     private final PartyService partyService;
     private final CurrentMember currentMember;
+    private final SocketSender socketSender;
 
     @Operation(summary = "채팅 메시지 전송", description = "채팅 메시지를 전송합니다.")
     @ApiResponse(responseCode = "200", description = "OK", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ChatSendResponseDto.class)))
@@ -58,7 +55,7 @@ public class MessageController {
         Message message = messageService.findById(messageId);
         Integer partyMemberNumber = partyService.countPartyMemberNumber(partyId);
         ChatSendResponseDto chatSendResponseDto = ChatSendResponseDto.createByMessage(message, partyMemberNumber, currentMember.id());
-        sendSocket(chatSendResponseDto, partyId);
+        socketSender.sendChatting(partyId, chatSendResponseDto);
         return DefaultResponse.res(StatusCode.OK, ChatResponseMessage.UPLOADED_MESSAGE, chatSendResponseDto);
     }
 
@@ -71,10 +68,5 @@ public class MessageController {
                                      @Schema(description = "이미지 파일을 업로드합니다.")
                                      @RequestPart MultipartFile imgFile) {
         return DefaultResponse.res(StatusCode.OK, ChatResponseMessage.UPLOADED_MESSAGE, new ChatMessageImageUploadResponseDto(""));
-    }
-
-    public void sendSocket(ChatSendResponseDto chatSendResponseDto, Long partyId){
-        SocketSendDto socketSendDto = SocketSendDto.createSocketSendDto(chatSendResponseDto);
-        simpMessagingTemplate.convertAndSend("/topic/chatting/" + partyId, socketSendDto);
     }
 }
